+++ evmrpc/block_test.go
@@ -1,172 +1,22 @@
 package evmrpc
 
 import (
-	"context"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"io"
 	"math/big"
 	"net/http"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/cosmos/cosmos-sdk/crypto/hd"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/sei-protocol/sei-chain/app"
-	"github.com/sei-protocol/sei-chain/x/evm/keeper"
 	"github.com/sei-protocol/sei-chain/x/evm/types"
 	"github.com/sei-protocol/sei-chain/x/evm/types/ethtx"
 	"github.com/stretchr/testify/require"
-	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/libs/bytes"
-	"github.com/tendermint/tendermint/libs/log"
-	"github.com/tendermint/tendermint/rpc/client/mock"
-	"github.com/tendermint/tendermint/rpc/coretypes"
-	tmtypes "github.com/tendermint/tendermint/types"
 )
-
-const BlockTestPort = 7779
-const BadBlockTestPort = 7780
-
-var EncodingConfig = app.MakeEncodingConfig()
-var TxConfig = EncodingConfig.TxConfig
-var Encoder = TxConfig.TxEncoder()
-var Decoder = TxConfig.TxDecoder()
-var Tx sdk.Tx
-
-type MockClient struct {
-	mock.Client
-}
-
-func (c *MockClient) mockBlock() *coretypes.ResultBlock {
-	return &coretypes.ResultBlock{
-		BlockID: tmtypes.BlockID{
-			Hash: bytes.HexBytes([]byte("0000000000000000000000000000000000000000000000000000000000000001")),
-		},
-		Block: &tmtypes.Block{
-			Header: tmtypes.Header{
-				ChainID:         "test",
-				Height:          8,
-				Time:            time.Unix(1696941649, 0),
-				DataHash:        bytes.HexBytes([]byte("0000000000000000000000000000000000000000000000000000000000000002")),
-				AppHash:         bytes.HexBytes([]byte("0000000000000000000000000000000000000000000000000000000000000003")),
-				LastResultsHash: bytes.HexBytes([]byte("0000000000000000000000000000000000000000000000000000000000000004")),
-				ProposerAddress: tmtypes.Address([]byte("0000000000000000000000000000000000000000000000000000000000000005")),
-				LastBlockID: tmtypes.BlockID{
-					Hash: bytes.HexBytes([]byte("0000000000000000000000000000000000000000000000000000000000000006")),
-				},
-			},
-			Data: tmtypes.Data{
-				Txs: []tmtypes.Tx{func() []byte {
-					bz, _ := Encoder(Tx)
-					return bz
-				}()},
-			},
-		},
-	}
-}
-
-func (c *MockClient) Genesis(context.Context) (*coretypes.ResultGenesis, error) {
-	return &coretypes.ResultGenesis{Genesis: &tmtypes.GenesisDoc{InitialHeight: 1}}, nil
-}
-
-func (c *MockClient) Block(ctx context.Context, height *int64) (*coretypes.ResultBlock, error) {
-	return c.mockBlock(), nil
-}
-
-func (c *MockClient) BlockByHash(ctx context.Context, hash bytes.HexBytes) (*coretypes.ResultBlock, error) {
-	return c.mockBlock(), nil
-}
-
-func (c *MockClient) BlockResults(ctx context.Context, height *int64) (*coretypes.ResultBlockResults, error) {
-	return &coretypes.ResultBlockResults{
-		TxsResults: []*abci.ExecTxResult{
-			{
-				Data: func() []byte {
-					bz, _ := Encoder(Tx)
-					return bz
-				}(),
-				GasWanted: 10,
-				GasUsed:   5,
-				Events: []abci.Event{{
-					Type: types.EventTypeEVMLog,
-					Attributes: []abci.EventAttribute{{
-						Key:   []byte(types.AttributeTypeContractAddress),
-						Value: []byte("0x1111111111111111111111111111111111111111111111111111111111111111"),
-					}, {
-						Key:   []byte(types.AttributeTypeBlockHash),
-						Value: []byte("0x1111111111111111111111111111111111111111111111111111111111111111"),
-					}, {
-						Key:   []byte(types.AttributeTypeBlockNumber),
-						Value: []byte("8"),
-					}, {
-						Key:   []byte(types.AttributeTypeData),
-						Value: []byte("xyz"),
-					}, {
-						Key:   []byte(types.AttributeTypeIndex),
-						Value: []byte("1"),
-					}, {
-						Key:   []byte(types.AttributeTypeTxIndex),
-						Value: []byte("2"),
-					}, {
-						Key:   []byte(types.AttributeTypeRemoved),
-						Value: []byte("true"),
-					}, {
-						Key:   []byte(types.AttributeTypeTopics),
-						Value: []byte("0x1111111111111111111111111111111111111111111111111111111111111111,0x1111111111111111111111111111111111111111111111111111111111111112"),
-					}, {
-						Key:   []byte(types.AttributeTypeTxHash),
-						Value: []byte("0x1111111111111111111111111111111111111111111111111111111111111113"),
-					}},
-				}},
-			},
-		},
-	}, nil
-}
-
-type MockBadClient struct {
-	MockClient
-}
-
-func (m *MockBadClient) Block(ctx context.Context, height *int64) (*coretypes.ResultBlock, error) {
-	return nil, errors.New("error block")
-}
-
-func (m *MockBadClient) BlockByHash(ctx context.Context, hash bytes.HexBytes) (*coretypes.ResultBlock, error) {
-	return nil, errors.New("error block")
-}
-
-func (m *MockBadClient) Genesis(context.Context) (*coretypes.ResultGenesis, error) {
-	return nil, errors.New("error genesis")
-}
-
-var EVMKeeper *keeper.Keeper
-var Ctx sdk.Context
-
-func init() {
-	EVMKeeper, _, Ctx = keeper.MockEVMKeeper()
-	httpServer, err := NewEVMHTTPServer(log.NewNopLogger(), TestAddr, BlockTestPort, rpc.DefaultHTTPTimeouts, &MockClient{}, EVMKeeper, func() sdk.Context { return Ctx }, Decoder)
-	if err != nil {
-		panic(err)
-	}
-	if err := httpServer.Start(); err != nil {
-		panic(err)
-	}
-	badHttpServer, err := NewEVMHTTPServer(log.NewNopLogger(), TestAddr, BadBlockTestPort, rpc.DefaultHTTPTimeouts, &MockBadClient{}, EVMKeeper, func() sdk.Context { return Ctx }, Decoder)
-	if err != nil {
-		panic(err)
-	}
-	if err := badHttpServer.Start(); err != nil {
-		panic(err)
-	}
-	time.Sleep(1 * time.Second)
-}
 
 func TestGetBlockByHash(t *testing.T) {
 	types.RegisterInterfaces(EncodingConfig.InterfaceRegistry)
@@ -205,7 +55,7 @@ func TestGetBlockByHash(t *testing.T) {
 	}))
 
 	body := "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockByHash\",\"params\":[\"0x0000000000000000000000000000000000000000000000000000000000000001\",true],\"id\":\"test\"}"
-	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BlockTestPort), strings.NewReader(body))
+	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 	require.Nil(t, err)
 	req.Header.Set("Content-Type", "application/json")
 	res, err := http.DefaultClient.Do(req)
@@ -253,7 +103,7 @@ func TestGetBlockByNumber(t *testing.T) {
 
 	for _, num := range []string{"0x8", "earliest", "latest", "pending", "finalized", "safe"} {
 		body := fmt.Sprintf("{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockByNumber\",\"params\":[\"%s\",true],\"id\":\"test\"}", num)
-		req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BlockTestPort), strings.NewReader(body))
+		req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 		require.Nil(t, err)
 		req.Header.Set("Content-Type", "application/json")
 		res, err := http.DefaultClient.Do(req)
@@ -264,7 +114,7 @@ func TestGetBlockByNumber(t *testing.T) {
 	}
 
 	body := "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockByNumber\",\"params\":[\"bad_num\",true],\"id\":\"test\"}"
-	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BlockTestPort), strings.NewReader(body))
+	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 	require.Nil(t, err)
 	req.Header.Set("Content-Type", "application/json")
 	res, err := http.DefaultClient.Do(req)
@@ -312,7 +162,7 @@ func TestGetBlockTransactionCount(t *testing.T) {
 	// get by block number
 	for _, num := range []string{"0x8", "earliest", "latest", "pending", "finalized", "safe"} {
 		body := fmt.Sprintf("{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockTransactionCountByNumber\",\"params\":[\"%s\"],\"id\":\"test\"}", num)
-		req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BlockTestPort), strings.NewReader(body))
+		req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 		require.Nil(t, err)
 		req.Header.Set("Content-Type", "application/json")
 		res, err := http.DefaultClient.Do(req)
@@ -324,7 +174,7 @@ func TestGetBlockTransactionCount(t *testing.T) {
 
 	// get error returns null
 	body := "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockTransactionCountByNumber\",\"params\":[\"0x8\"],\"id\":\"test\"}"
-	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BadBlockTestPort), strings.NewReader(body))
+	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestBadPort), strings.NewReader(body))
 	req.Header.Set("Content-Type", "application/json")
 	res, err := http.DefaultClient.Do(req)
 	require.Nil(t, err)
@@ -332,7 +182,7 @@ func TestGetBlockTransactionCount(t *testing.T) {
 	require.Nil(t, err)
 	require.Equal(t, "{\"jsonrpc\":\"2.0\",\"id\":\"test\",\"result\":null}\n", string(resBody))
 	body = "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockTransactionCountByNumber\",\"params\":[\"earliest\"],\"id\":\"test\"}"
-	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BadBlockTestPort), strings.NewReader(body))
+	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestBadPort), strings.NewReader(body))
 	req.Header.Set("Content-Type", "application/json")
 	res, err = http.DefaultClient.Do(req)
 	require.Nil(t, err)
@@ -340,7 +190,7 @@ func TestGetBlockTransactionCount(t *testing.T) {
 	require.Nil(t, err)
 	require.Equal(t, "{\"jsonrpc\":\"2.0\",\"id\":\"test\",\"result\":null}\n", string(resBody))
 	body = "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockTransactionCountByHash\",\"params\":[\"0x0000000000000000000000000000000000000000000000000000000000000001\"],\"id\":\"test\"}"
-	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BadBlockTestPort), strings.NewReader(body))
+	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestBadPort), strings.NewReader(body))
 	req.Header.Set("Content-Type", "application/json")
 	res, err = http.DefaultClient.Do(req)
 	require.Nil(t, err)
@@ -350,7 +200,7 @@ func TestGetBlockTransactionCount(t *testing.T) {
 
 	// get by hash
 	body = "{\"jsonrpc\": \"2.0\",\"method\": \"eth_getBlockTransactionCountByHash\",\"params\":[\"0x0000000000000000000000000000000000000000000000000000000000000001\"],\"id\":\"test\"}"
-	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, BlockTestPort), strings.NewReader(body))
+	req, err = http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 	require.Nil(t, err)
 	req.Header.Set("Content-Type", "application/json")
 	res, err = http.DefaultClient.Do(req)
