+++ evmrpc/echo_test.go
@@ -6,25 +6,13 @@ import (
 	"net/http"
 	"strings"
 	"testing"
-	"time"
 
-	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/require"
-	"github.com/tendermint/tendermint/libs/log"
 )
-
-const TestAddr = "127.0.0.1"
-const TestPort = 7777
-const TestWSPort = 7778
 
 func TestEcho(t *testing.T) {
 	// Test HTTP server
-	httpServer, err := NewEVMHTTPServer(log.NewNopLogger(), TestAddr, TestPort, rpc.DefaultHTTPTimeouts, nil, nil, nil, nil)
-	require.Nil(t, err)
-	require.Nil(t, httpServer.Start())
-	// wait for a second in case the actual server goroutine isn't ready yet
-	time.Sleep(1 * time.Second)
 	body := "{\"jsonrpc\": \"2.0\",\"method\": \"echo_echo\",\"params\":[\"something\"],\"id\":\"test\"}"
 	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("http://%s:%d", TestAddr, TestPort), strings.NewReader(body))
 	require.Nil(t, err)
@@ -36,11 +24,6 @@ func TestEcho(t *testing.T) {
 	require.Equal(t, "{\"jsonrpc\":\"2.0\",\"id\":\"test\",\"result\":\"something\"}\n", string(resBody))
 
 	// Test WS server
-	wsServer, err := NewEVMWebSocketServer(log.NewNopLogger(), TestAddr, TestWSPort, []string{"localhost"}, rpc.DefaultHTTPTimeouts)
-	require.Nil(t, err)
-	require.Nil(t, wsServer.Start())
-	// wait for a second in case the actual server goroutine isn't ready yet
-	time.Sleep(1 * time.Second)
 	headers := make(http.Header)
 	headers.Set("Origin", "localhost")
 	headers.Set("Content-Type", "application/json")
